refactor: 拆分使用说明界面hooks和配置文件以及工具类
